refactor(discord): split handlers.ts into handler modules
handlers.ts was 808 lines mixing chat registry, access control, queue,
attachment download, and the message pipeline, all sharing module-level
state. Split into a handlers/ folder mirroring the telegram layout:
  - state    — shared maps/sets (chat registry, access config, queues,
               rate windows, cooldowns)
  - registry — chatId ↔ Discord channel info accessors
  - access   — DM/guild/channel allowlists, admin, mention gating,
               unauthorized handling, rate limiting
  - context  — sender/reply context, attachment download/classify, chunked send
  - delivery — processAndReply pipeline
  - queue    — per-chat debounce queue
  - messages — the messageCreate handler

index.ts re-exports the identical public surface (incl. sendChunked +
DiscordChatInfo); all importers (discord/* + commands/* subdir) updated
to the barrel path.
